Rebasing onto latest main and fixing rebase issues
