+++ tests/tt_metal/tt_metal/CMakeLists.txt
@@ -64,7 +64,6 @@ add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/api)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/data_movement)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/debug_tools)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/device)
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/dm)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/dispatch)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/eth)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/integration)
@@ -86,7 +85,6 @@ add_custom_target(
         unit_tests_debug_tools_wormhole_b0
         unit_tests_debug_tools_blackhole
         unit_tests_device
-        unit_tests_dm
         unit_tests_dispatch
         unit_tests_eth_wormhole_b0
         unit_tests_eth_blackhole
