+++ tests/tt_metal/tt_metal/data_movement/CMakeLists.txt
@@ -1,4 +1,7 @@
-set(UNIT_TESTS_DATA_MOVEMENT_SRC ${CMAKE_CURRENT_SOURCE_DIR}/dram_unary/test_unary_dram.cpp)
+set(UNIT_TESTS_DATA_MOVEMENT_SRC
+    ${CMAKE_CURRENT_SOURCE_DIR}/dram_unary/test_unary_dram.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/one_to_one/test_one_to_one.cpp
+)
 
 add_executable(unit_tests_data_movement ${UNIT_TESTS_DATA_MOVEMENT_SRC})
 TT_ENABLE_UNITY_BUILD(unit_tests_data_movement)
