+++ tests/tt_metal/tt_metal/data_movement/dram_unary/kernels/writer_unary.cpp
@@ -16,22 +16,25 @@ void kernel_main() {
     constexpr uint32_t test_id = get_compile_time_arg_val(6);
 
     constexpr uint32_t transaction_size_bytes = transaction_num_pages * page_size_bytes;
+    constexpr uint32_t total_num_pages = num_of_transactions * transaction_num_pages;
 
     DeviceTimestampedData("Number of transactions", num_of_transactions);
     DeviceTimestampedData("Transaction size in bytes", transaction_size_bytes);
     DeviceTimestampedData("Test id", test_id);
 
-    for (uint32_t i = 0; i < num_of_transactions; i++) {
-        // TODO: Change dst address to change DRAM/core locations (single/multiple core)
-        uint64_t dst_noc_addr = get_noc_addr_from_bank_id<true>(bank_id, dst_addr);
+    cb_wait_front(cb_id_out0, total_num_pages);
+    {
+        DeviceZoneScopedN("WRITER");
+        for (uint32_t i = 0; i < num_of_transactions; i++) {
+            // TODO: Change dst address to change DRAM/core locations (single/multiple core)
+            uint64_t dst_noc_addr = get_noc_addr_from_bank_id<true>(bank_id, dst_addr);
+            uint32_t l1_read_addr = get_read_ptr(cb_id_out0);
 
-        cb_wait_front(cb_id_out0, transaction_num_pages);
-        uint32_t l1_read_addr = get_read_ptr(cb_id_out0);
-        noc_async_write(l1_read_addr, dst_noc_addr, transaction_size_bytes);
+            noc_async_write(l1_read_addr, dst_noc_addr, transaction_size_bytes);
+            noc_async_write_barrier();
 
-        noc_async_write_barrier();
-
-        cb_pop_front(cb_id_out0, transaction_num_pages);
-        dst_addr += transaction_size_bytes;
+            cb_pop_front(cb_id_out0, transaction_num_pages);
+            dst_addr += transaction_size_bytes;
+        }
     }
 }
